--- task1/requirements.txt
@@ -1,4 +1,0 @@
-beautifulsoup4==4.12.2
-requests==2.31.0
-selenium==4.19.0
-webdriver-manager==4.0.1
